[feat](landing): consolidate some components
